test: :white_check_mark: Only reset the database after integration tests

+++ src/testing/setup.ts
@@ -1,18 +1,20 @@
 import { afterAll } from 'vitest';
 import prisma from '../client/instance';
 
-afterAll(async () => {
-    const deleteQuestions = prisma.quizQuestion.deleteMany();
-    const deleteQuizzes = prisma.quiz.deleteMany();
-    const deleteUsers = prisma.user.deleteMany();
+afterAll(async (suite) => {
+    if (suite.tasks.some((task) => task.name.startsWith('@Integration'))) {
+        const deleteQuestions = prisma.quizQuestion.deleteMany();
+        const deleteQuizzes = prisma.quiz.deleteMany();
+        const deleteUsers = prisma.user.deleteMany();
 
-    await prisma.$transaction([deleteUsers, deleteQuizzes, deleteQuestions]);
+        await prisma.$transaction([deleteUsers, deleteQuizzes, deleteQuestions]);
 
-    const resetQuestions = prisma.$queryRaw`ALTER SEQUENCE "QuizQuestion_id_seq" RESTART WITH 1;`;
-    const resetQuiz = prisma.$queryRaw`ALTER SEQUENCE "Quiz_id_seq" RESTART WITH 1;`;
-    const resetUsers = prisma.$queryRaw`ALTER SEQUENCE "User_id_seq" RESTART WITH 1;`;
+        const resetQuestions = prisma.$queryRaw`ALTER SEQUENCE "QuizQuestion_id_seq" RESTART WITH 1;`;
+        const resetQuiz = prisma.$queryRaw`ALTER SEQUENCE "Quiz_id_seq" RESTART WITH 1;`;
+        const resetUsers = prisma.$queryRaw`ALTER SEQUENCE "User_id_seq" RESTART WITH 1;`;
 
-    await prisma.$transaction([resetUsers, resetQuiz, resetQuestions]);
+        await prisma.$transaction([resetUsers, resetQuiz, resetQuestions]);
 
-    await prisma.$disconnect();
+        await prisma.$disconnect();
+    }
 });
